refactor(miniwob): share BrowserGym OpenEnv flow
