+++ evalscope/api/benchmark/adapters/__init__.py
@@ -1,5 +1,6 @@
 from .agent_adapter import AgentAdapter, AgentLoopAdapter
 from .audio_language_adapter import AudioLanguageAdapter
+from .browsergym_adapter import BrowserGymOpenEnvAdapter
 from .default_data_adapter import DefaultDataAdapter
 from .function_call_adapter import FunctionCallAdapter
 from .image_edit_adapter import ImageEditAdapter
