+++ pyproject.toml
@@ -52,7 +52,6 @@ perf             = {file = ["requirements/perf.txt"]}
 app              = {file = ["requirements/app.txt"]}
 aigc             = {file = ["requirements/aigc.txt"]}
 sandbox          = {file = ["requirements/sandbox.txt"]}
-miniwob          = {file = ["requirements/miniwob.txt"]}
 service          = {file = ["requirements/service.txt", "requirements/perf.txt"]}
 mcp              = {file = ["requirements/mcp.txt"]}
 dev              = {file = ["requirements/dev.txt"]}
@@ -61,6 +60,7 @@ docs             = {file = ["requirements/docs.txt"]}
 ocr_bench        = {file = ["evalscope/benchmarks/ocr_bench/requirements.txt"]}
 omnidoc_bench    = {file = ["evalscope/benchmarks/omnidoc_bench/requirements.txt"]}
 maritime_ocr_bench = {file = ["evalscope/benchmarks/maritime_ocr_bench/requirements.txt"]}
+miniwob          = {file = ["evalscope/benchmarks/miniwob/requirements.txt"]}
 ifeval           = {file = ["evalscope/benchmarks/ifeval/requirements.txt"]}
 ifbench          = {file = ["evalscope/benchmarks/ifbench/requirements.txt"]}
 multi_if         = {file = ["evalscope/benchmarks/multi_if/requirements.txt"]}
